feat: add i18n support and category translation for VoluntaryOrganizations
