+++ src/common/i18n/locales/ar/common.json
@@ -21,6 +21,7 @@
   "ADVISORS_DESCRIPTION": "وراء كل مبادرة عظيمة فريق من المستشارين يقدمون الحكمة والتوجيه والدعم. تعرف على الخبراء الذين يرشدون مهمة Saayam For All.",
   "AGE": "العمر",
   "AI_ML_TEAM_LEAD": "قائد فريق الذكاء الاصطناعي وتعلم الآلة",
+  "AI_SUGGESTED": "مقترح بالذكاء الاصطناعي",
   "ALL": "الكل",
   "all live happily) guides our actions, driving": "يُلهِمنا هذا المبدأ ليقود أفعالنا نحو الخير، مما يعزز",
   "ALREADY_HAVE_ACCOUNT": "هل لديك حساب بالفعل؟",
@@ -75,6 +76,7 @@
   "CHAT_COMMUNITY_TOUCH_ORGANIZATIONS": "التواصل مع المجتمع والمنظمات",
   "CHAT_WITH_COMMUNITY_GET_IN_TOUCH": "الدردشة مع المجتمع والتواصل معنا",
   "CITY": "المدينة",
+  "CLEAR_SEARCH": "مسح البحث",
   "CLOSE": "إغلاق",
   "CLOSE_MENU": "إغلاق القائمة",
   "Clothing Support": "دعم الملابس",
@@ -195,6 +197,7 @@
   "Fill out the contact form and our team will reach out with onboarding steps!": "املأ نموذج الاتصال وسيتواصل فريقنا معك بخطوات الانضمام!",
   "FILL_CONTACT_FORM_TEAM_REACH": "املأ نموذج التواصل وسيتواصل فريقنا معك",
   "FILTER_BY": "تصفية حسب",
+  "FILTER_BY_SOURCE": "تصفية حسب المصدر:",
   "FIRST LANGUAGE PREFERENCE": "تفضيل اللغة الأولى",
   "First Name": "الاسم الأول",
   "FIRST_LANGUAGE_PREFERENCE": "تفضيل اللغة الأولى",
@@ -208,6 +211,7 @@
   "FOR_SELF": "لنفسي",
   "FOR_SELF_HELP_TEXT": "اختر 'نعم' إذا كنت تقدم هذا الطلب نيابة عن نفسك وإلا اختر 'لا' إذا كنت تطلب لشخص آخر.",
   "FORGOT_PASSWORD": "هل نسيت كلمة المرور؟",
+  "FOUND": "وُجدت",
   "FOUNDER_PRESIDENT_CEO_CTO": "المؤسس والرئيس والمدير التنفيذي ومدير التكنولوجيا",
   "Founder, President, CEO, CTO": "المؤسس والرئيس والمدير التنفيذي ومدير التكنولوجيا",
   "French": "الفرنسية",
@@ -223,6 +227,7 @@
   "Get Matching With a Volunteer": "الحصول على مطابقة مع متطوع",
   "GET_IN_TOUCH": "تواصل معنا",
   "GET_MATCHING_WITH_VOLUNTEER": "الحصول على مطابقة مع متطوع",
+  "GO_TO_DASHBOARD": "الذهاب إلى لوحة التحكم",
   "GOOGLE": "جوجل",
   "GOVERNING_LAW": "القانون الحاكم",
   "GOVERNING_LAW_CONTENT": "تخضع هذه الاتفاقية وتُفسر وفقاً للقوانين العالمية المطبقة على الأنشطة والتفاعلات التي تسهلها المنظمة. يجب رفع أي إجراءات قانونية ناشئة عن أو متعلقة بهذه الاتفاقية في المحاكم المناسبة للسلطة القضائية المتفق عليها بشكل متبادل من قبل كلا الطرفين.",
@@ -357,6 +362,10 @@
   "No Infrastructure or": "لا للبنية التحتية أو",
   "NO_BUILDING_SCHOOLS_HOSPITALS": "لا لبناء المدارس أو المستشفيات",
   "NO_INFRASTRUCTURE_MONEY_DISTRIBUTION": "لا للبنية التحتية أو توزيع الأموال",
+  "NO_MATCHING_RESULTS": "لا توجد نتائج مطابقة",
+  "NO_ORGANIZATIONS_AT_MOMENT": "لا توجد منظمات متاحة لهذه الفئة في الوقت الحالي.",
+  "NO_ORGANIZATIONS_FOUND": "لم يتم العثور على منظمات",
+  "NO_ORGANIZATIONS_MATCH": "لا توجد منظمات تطابق \"{{searchTerm}}\"",
   "NO_RESULTS": "لا توجد نتائج",
   "NON_PROFIT": "غير ربحية",
   "NONE_ACCOUNT": "ليس لديك حساب؟",
@@ -375,6 +384,7 @@
   "OPEN_TO_ALL_REGARDLESS": "مفتوح للجميع بغض النظر عن",
   "OPENS_NEW_WINDOW": "يفتح نافذة جديدة",
   "OR_WITH": "أو باستخدام",
+  "ORGANIZATION": "منظمة",
   "ORGANIZATION_DETAILS": "تفاصيل المنظمة",
   "ORGANIZATION_NAME": "اسم المنظمة",
   "ORGANIZATION_TYPE": "نوع المنظمة",
@@ -474,14 +484,19 @@
   "Receive help for their requests": "تلقي المساعدة لطلباتهم",
   "RECEIVE_ASSISTANCE": "تلقي المساعدة",
   "RECEIVE_HELP_REQUESTS": "استقبال طلبات المساعدة",
+  "RECOMMENDED_BY_AI": "موصى به من الذكاء الاصطناعي بناءً على طلبك",
   "Red Cross volunteers and staff work to deliver vital services – from providing relief and support to those in crisis.": "يعمل متطوعو وموظفو الصليب الأحمر على تقديم خدمات أساسية — من تقديم الإغاثة والدعم لمن هم في أزمة.",
   "REFRESH_PAGE": "تحديث الصفحة",
+  "REGISTERED": "مسجل",
+  "REGISTERED_WITH_US": "مسجل لدينا",
   "religion, location, country, or politics,": "الدين أو الموقع أو البلد أو السياسة،",
   "REMOTE": "عن بعد",
   "Request for Help": "طلب المساعدة",
   "Request Priority": "أولوية الطلب",
   "REQUEST_CATEGORY": "فئة الطلب",
   "REQUEST_DETAILS": "طَلَب تَفاصيل ",
+  "REQUEST_DETAILS_REQUIRED": "تفاصيل الطلب مطلوبة",
+  "REQUEST_DETAILS_REQUIRED_DESC": "يرجى الوصول إلى هذه الصفحة من طلب محدد لعرض المنظمات ذات الصلة.",
   "REQUEST_FOR_HELP": "طلب المساعدة",
   "REQUEST_TYPE": "نوع الطلب",
   "REQUEST_TYPE_HELP_TEXT": "حدد كيف تريد تقديم المساعدة: \"عن بعد\" للدعم الافتراضي أو \"شخصياً\" للمساعدة في الموقع.",
@@ -502,6 +517,7 @@
   "SCAN_ON_DONATE": "يمكنك المسح الضوئي أو النقر على التبرع",
   "SCAN_QR_CODE": "امسح رمز QR للتبرع",
   "SEARCH": "بحث",
+  "SEARCH_ORGANIZATIONS_PLACEHOLDER": "ابحث عن المنظمات بالاسم أو الموقع أو التواصل...",
   "SEARCH_PLACEHOLDER": "ابحث هنا…",
   "SECOND LANGUAGE PREFERENCE": "تفضيل اللغة الثانية",
   "SECOND_LANGUAGE_PREFERENCE": "تفضيل اللغة الثانية",
@@ -525,6 +541,7 @@
   "Sharing knowledge expands one's own understanding and the knowledge of others.": "مشاركة المعرفة توسع فهم الفرد ومعرفة الآخرين.",
   "SHARING_KNOWLEDGE_EXPANDS": "مشاركة المعرفة توسع الفهم",
   "shelter, education,": "المأوى والتعليم،",
+  "SHOWING_ORGANIZATIONS_FOR": "عرض المنظمات التي يمكنها المساعدة في",
   "SIGN_IN": "تسجيل الدخول",
   "SIGN_IN_TO_CONTINUE": "سجّل الدخول للمتابعة",
   "SIGN_OFF": "تسجيل الخروج",
@@ -608,6 +625,7 @@
   "TRAINING_SUPERVISION": "التدريب والإشراف",
   "Training_SUPERVISION_CONTENT": "ستوفر المنظمة التدريب والإشراف اللازمين للمتطوع. يوافق المتطوع على: <1/> - حضور جميع جلسات التدريب المطلوبة. <1/> - الالتزام بالتعليمات والإرشادات المقدمة من مشرفي المنظمة أو الموظفين. <1/> - طلب التوضيح أو المساعدة عند الحاجة لأداء واجباتهم بفعالية.",
   "TRAINING_SUPERVISION_CONTENT": "التدريب والإشراف والمحتوى",
+  "TRY_AGAIN": "حاول مرة أخرى",
   "TYPE": "نوع",
   "TYPE_YOUR_MESSAGE": "اكتب رسالتك",
   "U.S based nonprofit organization which provides a national digital infrastructure to serve volunteers and nonprofits organization.": "منظمة غير ربحية مقرها الولايات المتحدة توفر بنية تحتية رقمية وطنية لخدمة المتطوعين والمنظمات غير الربحية.",
@@ -623,6 +641,7 @@
   "VISION": "الرؤية",
   "VISION_DESCRIPTION": "تتصور Saayam For All عالماً يحصل فيه الجميع على المساعدة والموارد التي يحتاجونها للتغلب على العقبات وعيش حياة مُرضية. نطمح لأن نكون رواداً عالميين في المساعدات الإنسانية، مستفيدين من التكنولوجيا والتعاون لإنشاء شبكة دعم سلسة تتجاوز الحواجز الجغرافية والثقافية والاجتماعية والاقتصادية. رؤيتنا هي تنمية مجتمع حيث تُوجه أخلاقيات \"Manava Sevaye Madhava Seva\" (خدمة البشرية هي خدمة الله) و\"Sarve jana sukhino bhavantu\" (ليعش الجميع بسعادة) أفعالنا، ودفع التغيير الإيجابي وتعزيز عالم أكثر تعاطفاً وإنصافاً للجميع.",
   "VISION_STATEMENT": "بيان الرؤية",
+  "VISIT_WEBSITE": "زيارة الموقع",
   "Voluntary Organizations": "المنظمات التطوعية",
   "voluntary organizations without": "المنظمات التطوعية دون",
   "VOLUNTARY_ORGANIZATIONS": "المنظمات التطوعية",
