+++ src/pages/RequestDetails/VoluntaryOrganizations/components/LoadingSkeleton.jsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+const LoadingSkeleton = () => (
+  <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+    {[1, 2, 3, 4, 5, 6].map((i) => (
+      <div
+        key={i}
+        className="bg-white rounded-xl shadow-md border border-gray-100 p-6 animate-pulse"
+      >
+        <div className="flex items-start gap-4">
+          <div className="w-12 h-12 bg-gray-200 rounded-lg"></div>
+          <div className="flex-1">
+            <div className="h-5 bg-gray-200 rounded w-3/4 mb-2"></div>
+            <div className="h-4 bg-gray-200 rounded w-1/2"></div>
+          </div>
+        </div>
+        <div className="mt-4 space-y-3">
+          <div className="h-4 bg-gray-200 rounded w-full"></div>
+          <div className="h-4 bg-gray-200 rounded w-5/6"></div>
+          <div className="h-4 bg-gray-200 rounded w-4/6"></div>
+        </div>
+      </div>
+    ))}
+  </div>
+);
+
+export default LoadingSkeleton;
