+++ src/pages/RequestDetails/VoluntaryOrganizations/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useOrganizations } from "./useOrganizations";
