+++ src/pages/RequestDetails/VoluntaryOrganizations/utils/index.js
@@ -1,0 +1,8 @@
+export {
+  parseOrganizationsResponse,
+  getSourceType,
+  extractCategoryString,
+  filterBySource,
+  filterBySearchTerm,
+  countBySource,
+} from "./organizationHelpers";
